+++ src/Funds/FundForm/FundForm.js
@@ -390,6 +390,7 @@ const FundForm = ({
                         >
                           <FieldMultiSelection
                             dataOptions={fundOptions}
+                            id="fund-transfer-from"
                             label={<FormattedMessage id="ui-finance.fund.information.transferFrom" />}
                             name="fund.allocatedFromIds"
                             parse={parseMultiSelectionValue}
@@ -405,6 +406,7 @@ const FundForm = ({
                         >
                           <FieldMultiSelection
                             dataOptions={fundOptions}
+                            id="fund-transfer-to"
                             label={<FormattedMessage id="ui-finance.fund.information.transferTo" />}
                             name="fund.allocatedToIds"
                             parse={parseMultiSelectionValue}